+++ jib-core/src/main/java/com/google/cloud/tools/jib/docker/ImageTarball.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.jib.docker;
 
 import com.google.cloud.tools.jib.api.ImageReference;
+import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.docker.json.DockerManifestEntryTemplate;
 import com.google.cloud.tools.jib.image.Image;
 import com.google.cloud.tools.jib.image.Layer;
@@ -41,17 +42,17 @@ public class ImageTarball {
   private static final String LAYER_FILE_EXTENSION = ".tar.gz";
 
   private final Image image;
-  private final ImageReference imageReference;
+  private final BuildConfiguration buildConfiguration;
 
   /**
    * Instantiate with an {@link Image}.
    *
    * @param image the image to convert into a tarball
-   * @param imageReference image reference to set in the manifest
+   * @param buildConfiguration used for setting target image information in manifest
    */
-  public ImageTarball(Image image, ImageReference imageReference) {
+  public ImageTarball(Image image, BuildConfiguration buildConfiguration) {
     this.image = image;
-    this.imageReference = imageReference;
+    this.buildConfiguration = buildConfiguration;
   }
 
   public void writeTo(OutputStream out) throws IOException {
@@ -75,7 +76,10 @@ public class ImageTarball {
         CONTAINER_CONFIGURATION_JSON_FILE_NAME);
 
     // Adds the manifest to tarball.
-    manifestTemplate.setRepoTags(imageReference.toStringWithTag());
+    ImageReference imageReference = buildConfiguration.getTargetImageConfiguration().getImage();
+    for (String tag : buildConfiguration.getAllTargetImageTags()) {
+      manifestTemplate.addRepoTag(imageReference.withTag(tag).toStringWithTag());
+    }
     tarStreamBuilder.addByteEntry(
         JsonTemplateMapper.toByteArray(Collections.singletonList(manifestTemplate)),
         MANIFEST_JSON_FILE_NAME);