+++ jib-core/src/test/java/com/google/cloud/tools/jib/docker/ImageTarballTest.java
@@ -21,6 +21,8 @@ import com.google.cloud.tools.jib.api.ImageReference;
 import com.google.cloud.tools.jib.api.InvalidImageReferenceException;
 import com.google.cloud.tools.jib.blob.BlobDescriptor;
 import com.google.cloud.tools.jib.blob.Blobs;
+import com.google.cloud.tools.jib.configuration.BuildConfiguration;
+import com.google.cloud.tools.jib.configuration.ImageConfiguration;
 import com.google.cloud.tools.jib.docker.json.DockerManifestEntryTemplate;
 import com.google.cloud.tools.jib.image.Image;
 import com.google.cloud.tools.jib.image.Layer;
@@ -28,6 +30,8 @@ import com.google.cloud.tools.jib.image.LayerPropertyNotFoundException;
 import com.google.cloud.tools.jib.image.json.ContainerConfigurationTemplate;
 import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
 import com.google.cloud.tools.jib.json.JsonTemplateMapper;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import com.google.common.io.CharStreams;
 import com.google.common.io.Resources;
 import java.io.ByteArrayInputStream;
@@ -53,6 +57,8 @@ import org.mockito.junit.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class ImageTarballTest {
 
+  @Mock private BuildConfiguration mockBuildConfiguration;
+  @Mock private ImageConfiguration mockTargetImageConfiguration;
   @Mock private Layer mockLayer1;
   @Mock private Layer mockLayer2;
 
@@ -82,8 +88,14 @@ public class ImageTarballTest {
     Mockito.when(mockLayer2.getDiffId()).thenReturn(fakeDigestB);
     Image testImage =
         Image.builder(V22ManifestTemplate.class).addLayer(mockLayer1).addLayer(mockLayer2).build();
+    Mockito.when(mockBuildConfiguration.getTargetImageConfiguration())
+        .thenReturn(mockTargetImageConfiguration);
+    Mockito.when(mockBuildConfiguration.getAllTargetImageTags())
+        .thenReturn(ImmutableSet.of("tag", "another-tag", "tag3"));
+    Mockito.when(mockTargetImageConfiguration.getImage())
+        .thenReturn(ImageReference.parse("my/image"));
 
-    ImageTarball imageToTarball = new ImageTarball(testImage, ImageReference.parse("my/image:tag"));
+    ImageTarball imageToTarball = new ImageTarball(testImage, mockBuildConfiguration);
 
     ByteArrayOutputStream out = new ByteArrayOutputStream();
     imageToTarball.writeTo(out);
@@ -120,7 +132,11 @@ public class ImageTarballTest {
       String manifestJson =
           CharStreams.toString(
               new InputStreamReader(tarArchiveInputStream, StandardCharsets.UTF_8));
-      JsonTemplateMapper.readListOfJson(manifestJson, DockerManifestEntryTemplate.class);
+      DockerManifestEntryTemplate manifest =
+          JsonTemplateMapper.readListOfJson(manifestJson, DockerManifestEntryTemplate.class).get(0);
+      Assert.assertEquals(
+          ImmutableList.of("my/image:tag", "my/image:another-tag", "my/image:tag3"),
+          manifest.getRepoTags());
     }
   }
 }